update request headers and client properties
Updates the headers we send (incl.X-Super-Properties) as well as the
properties that are sent when IDENTIFYing to the gateway so that they
more closely align with current first-party client behavior.

Headers were made to live on the Session struct since the addition of
the heartbeat session ID, launch signature, and launch ID make it
untenable to keep this as a global in case of a multitenant scenario.